feat(cache): 实现语义缓存功能
- 新增 cache 包实现语义缓存存储逻辑，集成 Milvus 向量检索和 MySQL 数据存储
- 在 bootstrap 中注册语义缓存服务并将其加入对话流程 stages
- 修改 schema 配置文件名从 milvus_schema.json 到 document_chunk_collection.json
- 修正数据库表中 deleted 字段注释说明，统一为 0=未删除 1=已删除
- 添加 chat_cache_entry 表用于存储完整的语义缓存条目数据
- 更新配置结构体 SemanticCacheConf 添加 Collection 和修改 TTL 字段
- 在 conversation 包中添加 CacheWriteStage 和 SemanticCacheStage 实现缓存读写逻辑
- 修改实体模型 ChatCacheEntry 结构以支持新的缓存作用域字段
- 优化检索通道参数传递方式，在 keyword channel 中移除冗余参数

+++ internal/domain/chat/logic/conversation/context.go
@@ -264,7 +264,7 @@ type semanticCacheCtx struct {
 	hit        bool
 	entry      *entity.ChatCacheEntry
 	strategy   enum.ReuseStrategy
-	similarity float64
+	similarity float32
 }
 
 // MarkCacheHit 记录命中：挂载缓存条目与相似度
@@ -307,7 +307,7 @@ func (c *semanticCacheCtx) ReuseStrategy() enum.ReuseStrategy {
 }
 
 // CacheSimilarity 命中相似度（埋点用）
-func (c *semanticCacheCtx) CacheSimilarity() float64 {
+func (c *semanticCacheCtx) CacheSimilarity() float32 {
 	if c == nil {
 		return 0
 	}
